\name{sample.dist.R}
\alias{sample.dist.R}
\title{Sample from a probability distirbution using R}
\usage{
  sample.dist.R(prior, n)
}
\arguments{
  \item{prior}{data.frame with distn, parama, paramb}

  \item{n}{number of samples to return}
}
\value{
  vector with n random samples from prior
}
\description{
  Take n random samples from prior
}
\seealso{
  \link{pr.samp}
}

